remote changes

+++ devfile.cpp
@@ -4,8 +4,8 @@
 
 int main(int argc, char ** argv)
 {
-    int i = 0;
-
+    long i = 0; // changes on remote
+    
     for (int i = argc; i >= 0; i--)
     {
         printf("\n");
